fastapi/requirements.txt
nltk
fastapi
uvicorn
pydantic
mysql-connector-python
scikit-learn
